+++ Assets/Prefabs/Revolute Parts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63457b05dd7183943bba7aca2820ce02
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
