+++ Assets/Prefabs/Revolute Parts/logo-2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a512f1f2a066cf7458ca7928f18d3ebd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
