+++ Assets/Prefabs/Revolute Parts/logo-4.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f3a5e374a77295c48939b23c24747de8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
