+++ Assets/Prefabs/Revoluter 42.prefab
@@ -1,0 +1,186 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &4020571102471537478
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 4020571102471537529}
+  - component: {fileID: 4020571102471537531}
+  - component: {fileID: 4020571102471537528}
+  m_Layer: 0
+  m_Name: Revoluter 42
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4020571102471537529
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4020571102471537478}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: -7}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children:
+  - {fileID: 3477141066415273512}
+  - {fileID: 3477141067364712828}
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &4020571102471537531
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4020571102471537478}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 828bdbd4fa0ff53438360a10a05928b8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!65 &4020571102471537528
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4020571102471537478}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Size: {x: 3, y: 3, z: 3}
+  m_Center: {x: 0, y: 0, z: 0}
+--- !u!1001 &5057984402413981906
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 4020571102471537529}
+    m_Modifications:
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.8
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.23
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: -0.061048422
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0.99813485
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 187
+      objectReference: {fileID: 0}
+    - target: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 9023819233361398548, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+      propertyPath: m_Name
+      value: logo-2
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+--- !u!4 &3477141067364712828 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 8534562519822913966, guid: a512f1f2a066cf7458ca7928f18d3ebd, type: 3}
+  m_PrefabInstance: {fileID: 5057984402413981906}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &7057614760977050653
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 4020571102471537529}
+    m_Modifications:
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0.7
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 180
+      objectReference: {fileID: 0}
+    - target: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6555894119179059343, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+      propertyPath: m_Name
+      value: logo-4
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+--- !u!4 &3477141066415273512 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 5886457995315763765, guid: f3a5e374a77295c48939b23c24747de8, type: 3}
+  m_PrefabInstance: {fileID: 7057614760977050653}
+  m_PrefabAsset: {fileID: 0}
