+++ Assets/Prefabs/Revoluter 42.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c6f959cedb8f8b247b3bc52b9341955f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
