+++ Assets/Prefabs/Revoluter Globe.prefab
@@ -1,0 +1,183 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &8702611308606329138
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 8702611308606329137}
+  m_Layer: 0
+  m_Name: Revoluter Globe
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &8702611308606329137
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8702611308606329138}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: -7}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children:
+  - {fileID: 3483290484236908227}
+  - {fileID: 7634478240260952932}
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &8702611306952818972
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 8702611308606329137}
+    m_Modifications:
+    - target: {fileID: 4888956952064339301, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_Name
+      value: globe-earth
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.3
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.7
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+--- !u!4 &3483290484236908227 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 5229454174550816735, guid: b45166795e2254f57b899a9f754dd4d7, type: 3}
+  m_PrefabInstance: {fileID: 8702611306952818972}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &8702611308562909956
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 8702611308606329137}
+    m_Modifications:
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0.3
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.7
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1908734380221973210, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+      propertyPath: m_Name
+      value: globe_base
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+--- !u!4 &7634478240260952932 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 1240459547332876384, guid: 64a016c24981e0b4591df082dad4cbe1, type: 3}
+  m_PrefabInstance: {fileID: 8702611308562909956}
+  m_PrefabAsset: {fileID: 0}
