+++ Assets/Prefabs/Revoluter Globe.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 102bc3ef9b69318418cac82ef4f6e458
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
