+++ Assets/Scripts/Rotate.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3550e4e9bb604be4bb406c6527fee874
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
