+++ Assets/Scripts/Rotate/RevoluteRotate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 828bdbd4fa0ff53438360a10a05928b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
